Bind to interface support: Add new option 'ip'
The valid options to gen_tcp:listen, include an 'ip'
parameter which allows to specify which interface ip the
instance should listen on.

Implement a new option {ip, IpAddress} which can be passed
to misultin start_link and will eventually be passed to
gen_tcp:listen's options.

This particualar implementation only allows ip address strings
and not the tuple format {X,X,X,X} that erlang allows. I do
not see the advantage of exposing the tuple format to users
of misultin (in the spirit of principle of least suprise)

Signed-off-by: Essien Ita Essien <essiene@gmail.com>